Fix long prompt handling in mgsm
When an mgsm prompt with few-shot examples is too long, the number of
examples is reduced until it fits. However, due to the structure of the
base eval loop, this was done a little oddly - after the fewshot prompt
was built, the string was split instead of using the original structured
input data. This turned out to simply not work depending on the prompt
format. Specifically, for 0.3 or 0.4 prompts, the supposed delimiter for
few-shot instructions wasn't present, so the string wouldn't be split,
and the length couldn't be reduced. As a result, the code would give an
error that the 0-shot prompt (the prompt without any fewshot examples)
was already too long, even though that wasn't the case.

This changes the code to build the few-shot prompt using the structured
data inputs, and has been tested to work.

No other task uses a structure like this, but it should probably be
considered to avoid similar issues in the future. It's a little strange
because it requires overriding the `fewshot_context` function, but only
in order to wrap the original functionality with a length check.