Handle complicated prompts
This introduces a style for handling complex prompts and specifically
handles the case of JSLM Beta. This is handled by using a function that
takes the name of task as input. This allows for full customization
without requiring details specification when actually running an eval
suite.

The style is simple - instead of mapping to a numeric version like 0.2,
a shortname for a prompt can map to a callable that takes the task name.
This allows for any kind of custom logic.

This may not be the simplest or best approach, but it required few
changes, keeps everything in one place, and touches nothing else in the
code base, so it should be easy to change later if necessary.